refactor(ui): move ui-header to top-level include
